refactor multiclass warnings into separate class
